requirements.txt
numpy~=2.2.4
ai2thor~=5.0.0
torch~=2.7.0
torchvision~=0.22.0
tqdm~=4.67.1
torch-geometric~=2.6.1
matplotlib~=3.10.3
pillow~=11.2.1
optuna~=4.4.0
tensorboard~=2.19.0
pandas~=2.3.1
timm~=1.0.22
opencv-python~=4.12.0.88
lmdb~=1.5.1
pymysql~=1.1.2